refactor(arcane-infra): extract NodeCore from run_node_loop (#179) (#185)
* refactor(arcane-infra): extract NodeCore from run_node_loop (#179)

Extract an embeddable NodeCore from node_runner::run_node_loop — a pure,
behavior-preserving refactor. run_node_loop stays as a thin loop/sleep
driver over NodeCore. No behavior change; no API removed; binaries untouched.

- Created node_core.rs with NodeCore struct and NodeConfig
- NodeCore::new() does all setup (Redis, channels, I/O threads)
- NodeCore::tick() executes one loop iteration (drain, simulate, tick, broadcast)
- run_node_loop becomes a ~10-line driver over NodeCore
- Moved merge_with_neighbor_latest and its tests to node_core
- Both arcane_node.rs and arcane_rapier_node.rs remain completely unchanged


* fix: correct dt_seconds calculation to preserve integer-millisecond quantization

In NodeCore::new, use Duration::from_millis for precise millisecond
quantization, matching the original run_node_loop behavior exactly.


---------
